Create tiered-cart-discount-by-spend
Use this script to offer a dollar amount discount that increases according to the total value of the items in their cart. This discount is spread as evenly as possible across the individual items in the cart.

For example, offer customers $25 off their total if they spend more than $150, $50 off if they spend more than $300, or $75 off if they spend more than $400.